Optimize imports & Global Exception Handler & clean code backend

+++ BackEnd/src/main/java/tfg/backend/controllers/UserController.java
@@ -3,9 +3,13 @@ package tfg.backend.controllers;
 import lombok.Data;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import tfg.backend.models.Asignatura;
-import tfg.backend.models.Matricula;
 import tfg.backend.models.Usuario;
 import tfg.backend.services.interfaces.IAsignaturaService;
 import tfg.backend.services.interfaces.IUserService;
@@ -14,7 +18,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Data
-class UserWithRole{
+class UserWithRole {
     public Usuario user;
     public String rol;
 }
@@ -40,15 +44,16 @@ public class UserController {
     public Usuario crearUsuario(@RequestBody UserWithRole user) {
         //TODO reemplazar con contraseña aleatoria
         user.user.setPassword(bCryptPasswordEncoder.encode("password"));
-        return userService.create(user.user,user.rol);
+        return userService.create(user.user, user.rol);
     }
 
     @GetMapping("/profesores")
     public List<Usuario> getProfesores() {
         return userService.getProfesores();
     }
+
     @GetMapping("/profesores/{id}")
-    public Usuario getProfesoresID(@PathVariable("id") Integer id) {
+    public Usuario getProfesoresID(@PathVariable("id") Long id) {
         return userService.findById(id);
     }
 
@@ -61,18 +66,19 @@ public class UserController {
     public List<Usuario> getAlumnos() {
         return userService.getAlumnos();
     }
+
     @GetMapping("/alumnos/{id}")
-    public Usuario getAlumnosID(@PathVariable("id") Integer id) {
+    public Usuario getAlumnosID(@PathVariable("id") Long id) {
         return userService.findById(id);
     }
 
     @GetMapping("/profesores/{id}/asignaturas")
-    public Collection<Asignatura> getProfesoresIDAsignatura(@PathVariable("id") Integer id) {
+    public Collection<Asignatura> getProfesoresIDAsignatura(@PathVariable("id") Long id) {
         return userService.findById(id).getAsignaturasImpartidas();
     }
 
     @GetMapping("/alumnos/{id}/asignaturas")
-    public Collection<Asignatura> getMatriculasAlumno(@PathVariable("id") Integer id) {
+    public Collection<Asignatura> getMatriculasAlumno(@PathVariable("id") Long id) {
         return asignaturaService.getAsignaturasAlumno(id);
     }
 }
