+++ BackEnd/src/main/java/tfg/backend/services/UserService.java
@@ -31,7 +31,7 @@ public class UserService implements UserDetailsService, IUserService {
     public Usuario create(Usuario user, String role) {
         Role r = this.roleReposority.findByNombre(role).orElse(null);
         user.setRoles(Collections.singletonList(r));
-        Usuario usuario =  this.iUsuarioReposority.save(user);
+        Usuario usuario = this.iUsuarioReposority.save(user);
         return null;
     }
 
@@ -39,30 +39,32 @@ public class UserService implements UserDetailsService, IUserService {
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         Usuario user = iUsuarioReposority.findByUser(username).orElse(null);
 
-        if(user==null)
-            throw  new UsernameNotFoundException("Usuario no encontrado");
+        if (user == null)
+            throw new UsernameNotFoundException("Usuario no encontrado");
 
         List<GrantedAuthority> authorities = new ArrayList<>();
-        for(Role r:user.getRoles()) {
-        	authorities.add(new SimpleGrantedAuthority(r.getNombre()));
+        for (Role r : user.getRoles()) {
+            authorities.add(new SimpleGrantedAuthority(r.getNombre()));
         }
 
-        return new User(user.getUser(), user.getPassword(),authorities);
+        return new User(user.getUser(), user.getPassword(), authorities);
     }
 
-    public Usuario getUser(String username){
+    public Usuario getUser(String username) {
+
         return iUsuarioReposority.findByUser(username).orElse(null);
     }
 
-    public List<Usuario> all(){
+    public List<Usuario> all() {
+
         return iUsuarioReposority.findAll();
     }
 
-    public Usuario findById(Integer id){
+    public Usuario findById(Long id) {
         return this.iUsuarioReposority.findById(id).orElse(null);
     }
 
-    public List<Usuario> getProfesores(){
+    public List<Usuario> getProfesores() {
         Role r = new Role();
         r.setId(2l);
         r.setNombre("PROFESOR");
