+++ BackEnd/src/main/java/tfg/backend/services/interfaces/IUserService.java
@@ -9,14 +9,21 @@ import tfg.backend.models.Usuario;
 import java.util.List;
 
 @Service
-public interface IUserService extends UserDetailsService{
+public interface IUserService extends UserDetailsService {
 
     UserDetails loadUserByUsername(String username) throws UsernameNotFoundException;
+
     Usuario getUser(String username);
+
     List<Usuario> all();
-    Usuario create(Usuario user,String role);
-    Usuario findById(Integer id);
+
+    Usuario create(Usuario user, String role);
+
+    Usuario findById(Long id);
+
     List<Usuario> getProfesores();
+
     List<Usuario> getAlumnos();
+
     List<Usuario> getAdmins();
 }
